% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{More interesting pop-ups for the interactive map used with the eq_map() function}
\usage{
eq_create_label(mapdata)
}
\arguments{
\item{mapdata}{A cleaned data frame with data obtained from NOAA website}
}
\value{
An HTML label that can be used as the annotation text in the leaflet map.
}
\description{
More interesting pop-ups for the interactive map used with the eq_map() function
}
\details{
This function return a more interesting pop-ups for the interactive map created with the eq_map() function.
The function is called before the eq_map() function and used the same cleaned data.
The cleaning of the data frame is done with the eq_clean_data() and eq_location_clean() functions of this package.
For each earthquake, the pop-up will show its "Location", "Total deaths" and "Magnitude".
}
\examples{
\dontrun{
readr::read_delim("signif.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}

}
