% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning NOAA earthquake data location values}
\usage{
eq_location_clean(locationtoclean)
}
\arguments{
\item{locationtoclean}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned LOCATION_NAME column
}
\description{
Cleaning NOAA earthquake data location values
}
\details{
After downloading and reading in the dataset from NOAA site,
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
National Geophysical Data Center / World Data Service (NGDC/WDS): Significant Earthquake Database.
National Geophysical Data Center, NOAA. doi:10.7289/V5TD9V7K
The function cleans the LOCATION_NAME column by stripping out the country name (including the colon)
and converts names to title case (as opposed to all caps).
}
\examples{
\dontrun{
library(readr)
data <- readr::read_delim("signif.txt", delim = "\\t")
data <- eq_location_clean(data)
}

}
