% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timelinelabel}
\alias{geom_timelinelabel}
\title{Visualize the times at which earthquakes occur within certain countries}
\usage{
geom_timelinelabel(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to [layer()]. These are
  often aesthetics, used to set an aesthetic to a fixed value, like
  `colour = "red"` or `size = 3`. They may also be parameters
  to the paired geom/stat.

#' @section Aesthetics:
\code{geom_timeline_label} understands the following aesthetics (properties of the plot that can show certain elements of the data)
 (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}     #Time variable
  \item \strong{\code{label}} #Factor for adding annotations to the earthquake data
  \item \code{y}              #Factor indicating some stratification
  \item \code{n_max}          #number of earthquakes to subset according to their magnitude (EQ_PRIMARY)
  \item \code{y_length}       #vertical line length to each data point
}}
}
\value{
The geom \code{geom_timeline_label} used with the \code{ggplot} function and the \code{geom_timeline} geom,
add annotations to the n_max largest (by magnitude) earthquakes.
}
\description{
This geom adds a vertical line to each data point with a text annotation (e.g. the location of the earthquake) attached to each line.
There is also an option to subset to n_max number of earthquakes, where we take the n_max largest (by magnitude) earthquakes.
}
\details{
The data to be used with this geom must be downloaded and readed from NOAA site,
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
}
\examples{
# The data must be cleaned using the function \\code{eq_clean_data}, included in the package.
# The LOCATION_NAME colomn of the data must be cleaned using the function \\code{eq_location_clean},
# included in the package.
# Aesthetics can be specified in the \\code{ggplot} function or in \\code{geom_timeline} geom function
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\\t")
data <- eq_clean_data(data)
data <- eq_location_clean(data)
data \%>\%
dplyr::filter(COUNTRY == c("MEXICO","USA") & lubridate::year(DATE) >= 2010) \%>\%
ggplot(aes(x=DATE,y=COUNTRY,color=TOTAL_DEATHS,size=EQ_PRIMARY)) +
geom_timeline(alpha=.5) +
geom_timelinelabel(aes(label=LOCATION_NAME),n_max=3) +
theme(legend.position="bottom", legend.box="horizontal", plot.title=element_text(hjust=0.5)) +
ggtitle("Earthquakes Visualization Tool") +
labs(size = "Richter scale value", color = "# deaths")
}
}
