% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelinelabel}
\alias{GeomTimelinelabel}
\title{GeomTimelinelabel}
\format{An object of class \code{GeomTimelinelabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelinelabel
}
\description{
GeomTimelinelabel Geom coding
}
\keyword{datasets}
